refactor(backend): add comment headers, delete helpers
Almost all helpers were wrappers for commands.

+++ src-tauri/src/resources/structs.rs
@@ -1,3 +1,7 @@
+// File:     resources/structs.rs
+// Purpose:  Structs for the Resources tool.
+// Authors:  Ojos Project & Iris contributors
+// License:  GNU General Public License v3.0
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
@@ -7,5 +11,5 @@ pub struct Resource {
     pub url: String,
     pub organization: String,
     pub category: String,
-    pub last_updated: f32,
+    pub last_updated: i64,
 }
